ENH: Port AlgorithmResult from Quantopian.
Port the AlgorithmResult class from Quantopian's research
environment. This replaces the old backtest output structure, which was
a DataFrame containing a mix of scalars, dictionaries, and lists of
dictionaries.

`TradingAlgorithm.run` now returns a BacktestResult. Most of the changes
here are updating tests to read values out of this new format instead of
from the old one.